Create UserGroupComparison.ps1
Compares user groups to the list of groups in the OU description and provides a CSV output with all users, their OU, the missing groups, the extra groups, and the groups they are in. Separate file provided to remediate the user groups (RemediateUserGroups.ps1) - outputs to Reports folder and runs "dryrun" by default.